Removed NextStepController and fxml and rerouted welcomecomtroller to go into GymAppHomeController and fxml after signing into an account
