+++ src/main/resources/com/example/firebaseadminjavafx/gymapp-home.fxml
@@ -1,40 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.VBox?>
 
-<VBox xmlns="http://javafx.com/javafx/21"
-      xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="com.example.firebaseadminjavafx.GymAppHomeController"
-      alignment="CENTER" spacing="15.0"
-      prefWidth="400.0" prefHeight="400.0">
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
 
-    <children>
-        <Label text="Gym Home" style="-fx-font-size: 18px; -fx-font-weight: bold;" />
+<VBox alignment="CENTER" prefHeight="400.0" prefWidth="400.0" spacing="15.0" xmlns="http://javafx.com/javafx/17.0.12"
+      xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.firebaseadminjavafx.GymAppHomeController">
 
-        <Button fx:id="gymLayoutButton"
-                text="Gym Layout"
-                onAction="#handleGymLayout"
-                prefWidth="180" />
+    <HBox spacing="6" alignment="CENTER">
+        <Label text="Email:"/>
+        <Label fx:id="emailValue" text="(loading)"/>
+    </HBox>
 
-        <Button fx:id="workoutPlanningButton"
-                text="Workout Planning"
-                onAction="#handleWorkoutPlanning"
-                prefWidth="180" />
+    <Label style="-fx-font-size: 18px; -fx-font-weight: bold;" text="Gym Home"/>
 
-        <Button fx:id="progressTrackerButton"
-                text="Workout Progress Tracker"
-                onAction="#handleProgressTracker"
-                prefWidth="180" />
+    <Button fx:id="gymLayoutButton" onAction="#handleGymLayout" prefWidth="180" text="Gym Layout"/>
 
-        <Button fx:id="classSchedulerButton"
-                text="Classes"
-                onAction="#handleClassScheduling"
-                prefWidth="180" />
+    <Button fx:id="workoutPlanningButton" onAction="#handleWorkoutPlanning" prefWidth="180" text="Workout Planning"/>
 
-        <Button fx:id="backButton"
-                text="Back"
-                onAction="#handleBack"
-                prefWidth="120" />
-    </children>
+    <Button fx:id="progressTrackerButton" onAction="#handleProgressTracker" prefWidth="180"
+            text="Workout Progress Tracker"/>
+
+    <Button fx:id="classSchedulerButton" onAction="#handleClassScheduling" prefWidth="180" text="Classes"/>
+
+    <Button fx:id="myProfileButton" onAction="#handleMyProfile" prefWidth="180" text="My Profile"/>
+
+    <Button fx:id="returnToWelcomeButton" onAction="#handleReturnToWelcome" prefWidth="180" text="Return to Welcome"/>
 </VBox>
