--- src/main/resources/com/example/firebaseadminjavafx/next-step.fxml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-
-<VBox xmlns="http://javafx.com/javafx"
-      xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.firebaseadminjavafx.NextStepController"
-      alignment="CENTER"
-      spacing="15"
-      prefWidth="400"
-      prefHeight="400">
-
-    <padding>
-        <Insets top="25" right="25" bottom="25" left="25"/>
-    </padding>
-
-    <children>
-        <Label text="Next Step"
-               style="-fx-font-size: 20px; -fx-font-weight: bold;"/>
-
-        <HBox spacing="6" alignment="CENTER">
-            <Label text="Email:"/>
-            <Label fx:id="emailValue" text="(loading)"/>
-        </HBox>
-
-        <Button fx:id="goGymHomeButton"
-                text="Go to GymHome"
-                onAction="#handleGoGymHome"
-                prefWidth="180"/>
-
-        <Button fx:id="enterInfoButton"
-                text="Enter Info"
-                onAction="#handleEnterInfo"
-                prefWidth="180"/>
-
-        <Button fx:id="returnToWelcomeButton"
-                text="Return to Welcome"
-                onAction="#handleReturnToWelcome"
-                prefWidth="180"/>
-    </children>
-</VBox>
